nambahin form dan card buat pertandingan. fitur nambahin pertandingan dah oke, tapi styling ga tau kenapa belum applied. card dah oke, tinggal bikin tombol buat liat detail pertandingan sama mungkin highlightnya dipindahin ke halaman detail pertandingan aja. sama tampilan match.html nya kok malah ngawur dah
