+++ match/templates/add_match.html
@@ -1,0 +1,72 @@
+{% extends 'base_after_login.html' %} 
+
+{% block content %}
+<div class="content" style="position: relative;">
+
+    <!-- Tombol Back di pojok kiri atas -->
+    <div style="position: absolute; top: 10px; left: 20px;">
+        <a href="{% url 'match:show_match' %}">
+            <button style="
+                background-color: #a4161a;
+                color: white;
+                border: none;
+                padding: 8px 14px;
+                border-radius: 8px;
+                font-weight: 600;
+                cursor: pointer;
+                transition: 0.3s;
+            "
+            onmouseover="this.style.backgroundColor='#6e0f11'"
+            onmouseout="this.style.backgroundColor='#a4161a'">
+                ← Back to Merchandise List
+            </button>
+        </a>
+    </div>
+    
+    <!-- Judul -->
+    <h1 style="text-align: center; color: #a4161a; margin-top: 50px;">
+        Garuda Lounge Merchandise
+    </h1>
+      
+      <form method="POST" class="space-y-6">
+        {% csrf_token %}
+        {% for field in form %}
+          <div>
+            <label for="{{ field.id_for_label }}" class="block text-sm font-medium text-gray-700 mb-2">
+              {{ field.label }}
+            </label>
+            <div class="w-full">
+              {{ field }}
+            </div>
+            {% if field.help_text %}
+              <p class="mt-1 text-sm text-gray-500">{{ field.help_text }}</p>
+            {% endif %}
+            {% for error in field.errors %}
+              <p class="mt-1 text-sm text-red-600">{{ error }}</p>
+            {% endfor %}
+          </div>
+        {% endfor %}
+        
+        <div style="text-align: center; margin: 20px 0;">
+        <a href="{% url 'match:add_match' %}">
+            <button style="
+                background-color: #a4161a;
+                color: white;
+                border: none;
+                padding: 10px 16px;
+                border-radius: 8px;
+                font-weight: 600;
+                cursor: pointer;
+                transition: 0.3s;
+            "
+            onmouseover="this.style.backgroundColor='#6e0f11'"
+            onmouseout="this.style.backgroundColor='#a4161a'">
+                Simpan Pertandingan
+            </button>
+        </a>
+    </div>
+      </form>
+    </div>
+  </div>
+</div>
+{% endblock %}
